+++ contrib/storage-http/README.md
@@ -666,8 +666,9 @@ an API has parameters which are part of the URL AND these parameters are dynamic
 
 In this case, there are two functions `http_get_url` and `http_get` which you can use to faciliate these joins. 
 
-* `http_get('<storage_plugin_name>', <params>)`:  This function accepts a storage plugin as input and an optional list of parameters to include in a URL.
-* `http_get_url(<url>, <params>)`:  This function works in the same way except that it does not pull any configuration information from existing storage plugins.
+* `http_request('<storage_plugin_name>', <params>)`:  This function accepts a storage plugin as input and an optional list of parameters to include in a URL.
+* `http_get(<url>, <params>)`:  This function works in the same way except that it does not pull any configuration information from existing storage plugins.  The input url for 
+  the `http_get` function must be a valid URL. 
 
 ### Example Queries
 Let's say that you have a storage plugin called `github` with an endpoint called `repos` which points to the url: https://github.com/orgs/{org}/repos.  It is easy enough to 
@@ -681,7 +682,7 @@ WHERE org='apache'
 However, if you had a file with organizations and wanted to join this with the API, the query would fail.  Using the functions listed above you could get this data as follows:
 
 ```sql
-SELECT http_get('github.repos', `org`)
+SELECT http_request('github.repos', `org`)
 FROM dfs.`some_data.csvh`
 ```
 or
